GNUmakefile: improve existing build rules
First, we add .PHONY declarations for the phony targets. This gives a
slight performance improvement and avoids a potential conflict with
files whose name is the same as that of the target.

Second, we rename the main program targets, like poly -> poly.x. The
output of that target is a file named poly.x, and previously, "make
poly" would always remake it. Now the executable will not be rebuilt
unless one of its dependencies has changed (the usual "make"
behavior).

Finally, thanks to the second change, we are able to replace the
hard-coded outputs with the name of the target, e.g. poly.x -> $@.
This avoids duplication and will make it easier to add variants
later.