GNUmakefile: drop header prerequisites from the executable targets
Our executable targets depend on the corresponding object files, which
themselves depend on certain headers. As a result, the executables do
not need to depend on the headers again.

For example, poly.x depends on poly.o, and poly.o depends on both
Global.h and LG.h. If either of those headers is modified, then poly.o
will be rebuilt, triggering a rebuild of poly.x. Thus in the rule,

  poly.x: poly.o $(OBJECTS) LG.o Global.h LG.h

the prerequisites Global.h and LG.h are redundant. This commit removes
them.