GNUmakefile: use automatic $^ variable to refer to object files
Each executable target lists some object files as prerequisites; for
example,

  poly.x: poly.o $(OBJECTS) LG.o

The build rule that follows must then repeat those object files in
order to pass them to the compiler/linker:

  $(CC) ... -o $@ poly.o $(OBJECTS) LG.o

This can be simplified a bit because GNU Make provides an automatic
variable called $^ that lists all of the target's prerequisites. So,
in the example above, $^ is "poly.o $(OBJECTS) LG.o". In other words,
it's exactly what we want to pass to the compiler/linker.

This commit updates the five executable rules in this manner to avoid
repeating the names of the object files.