*.c: use foo(void) to indicate that foo takes no arguments
In C23, foo() means that foo takes an unknown number of arguments. The
way to indicate that foo takes zero arguments is instead foo(void).
Compilers will complain about this in the future; for now it can be
triggered with -Wstrict-prototypes.

We add "void" in a few places to future-proof the code base.