Rat.{c,h}: update Map_Permut prototype
The function Map_Permut takes a function argument called ArgFun,
both declared in Rat.h:

  ifdef __cplusplus
  #define ARG_FUN void (*ArgFun)(int *d,int *pi,int *pinv,void *info)
  #else
  #define ARG_FUN void ( ArgFun() )
  #endif
  void Map_Permut(int *d,int *pi,int *pinv,ARG_FUN,void *AuxPtr);

ArgFun is indeed eventually used as a function in Rat.c:

  for (j=*d;j>0;j--) {
    ...
  } (ArgFun)(d,pi,pinv,AuxPtr);

To me it looks like the __cplusplus branch declares the argument type
correctly, while the other branch attempts to declare it with no
arguments. In C23, the latter is actually an error -- it could be
fixed by using ArgFun(void) rather than ArgFun(). However... the
number and types of of its arguments are fixed and correct in the
__cplusplus branch, so why not just use it?

That's what this commit does. As a result, the build now succeeds with
CFLAGS="-Werror=strict-prototypes".