SingularInput.c: use file descriptors to pass around temporary files
This is phase one of a switch from tmpname() to mkstemp(). For now,
all we do is change one call from fopen() to open(), so that we now
obtain a file descriptor instead of a FILE pointer. That necessitates
changing several fprintf() calls to dprintf(), as well as updating
some of the function signatures in Mori.h and MoriCone.c.