SingularInput.c: move CatFile() from MoriCone.c
The CatFile() function is called from only one location, inside
SingularInput.c. Moreover, newer compilers (clang-17) complain
about the implicitly defined function because SingularInput.c
does not know that the function is defined in MoriCone.c. To fix
that, the simplest solution is to move the function into the
file that uses it, as opposed to (say) adding a prototype to
some header.