SingularInput.c: replace tmpnam() with mkstemp()
Newer, more secure functions like mkstemp() exist to replace the old
tmpnam(). Some toolchains even warn about this; for example,

  SingularInput.c:(.text+0x306): warning: the use of `tmpnam' is
  dangerous, better use `mkstemp'

This commit updates the HyperSurfSingular() function to use mkstemp()
instead of tmpnam() and open(). A minor change is required in CatFile()
as well to address the length of the resulting temporary file.