GNUmakefile: don't override POLY_Dmax when building prog.x
Previously, prog.x was copied from prog-6d.x after the latter was
built. This avoids an extra compilation, but accidentally broke one of
the documented ways of increasing POLY_Dmax: by editing Global.h, and
recompiling. This commit compiles the prog.x separately, and adds a
special case to one of the build rules so that -DPOLY_Dmax is omitted
when doing so.